expose watcher
