+++ core/pkg/server/stream.go
@@ -73,17 +73,18 @@ func NewStream(ctx context.Context, settings *service.Settings, streamId string)
 		outChan:      make(chan *service.ServerResponse, BufferSize),
 	}
 
+	watcher := watcher.New(watcher.WithLogger(s.logger))
 	s.handler = NewHandler(s.ctx, s.logger,
 		WithHandlerSettings(s.settings),
 		WithHandlerFwdChannel(make(chan *service.Record, BufferSize)),
 		WithHandlerOutChannel(make(chan *service.Result, BufferSize)),
 		WithHandlerSystemMonitor(monitor.NewSystemMonitor(s.logger, s.settings, s.loopBackChan)),
-		WithHandlerFileHandler(NewFilesHandler(s.logger, s.settings)),
-		WithHandlerTBHandler(NewTBHandler(s.logger, s.settings, s.loopBackChan)),
-		WithHandlerFileTransferHandler(NewFileTransferHandler()),
+		WithHandlerFileHandler(NewFilesHandler(watcher, s.logger, s.settings)),
+		WithHandlerTBHandler(NewTBHandler(watcher, s.logger, s.settings, s.loopBackChan)),
+		WithHandlerFilesInfoHandler(NewFilesInfoHandler()),
 		WithHandlerSummaryHandler(NewSummaryHandler(s.logger)),
 		WithHandlerMetricHandler(NewMetricHandler()),
-		WithHandlerWatcher(watcher.New(watcher.WithLogger(s.logger))),
+		WithHandlerWatcher(watcher),
 	)
 
 	s.writer = NewWriter(s.ctx, s.logger,
